Stack variables now work, except for frame offsets
A good portion of simple functions compile as intended, the main issue
is only with a high number of operations and -O0, and that frame
offests, when the function needs to push and pop registers, are wrong.
This is simple to solve (the solution will come in the next commit) as
AVR uses MachineFunctionInfo to do that already, so using a similar (but
simpler) method should work correctly.