Changing calling conv to compile complex functions
Now, arguments are implemented fully (for int16_t and uint16_t at least)
both via stack and registers. Because of the change in calling conv, now
it is possible to compile code without the fear of running out of
registers, because there are at least 3 not used for function arguments:
the necessary number of registers to push onto the stack (ICMC not having
a load/store indirect with SP plus offset really kills performance for
stack variables).